refine: add TabCleaner to TabOrganizerPage
