+++ src/view/features/options/pages/Settings/index.tsx
@@ -2,12 +2,10 @@ import Stack from "@mui/material/Stack";
 
 import PopupElementScaleSettingForm from "./PopupElementScaleSettingForm";
 import PopupSizeSettingForm from "./PopupSizeSettingForm";
-import TabCleanerSettingForm from "./TabCleanerSettingForm";
 
 const Settings = () => {
   return (
     <Stack spacing={2}>
-      <TabCleanerSettingForm />
       <PopupSizeSettingForm />
       <PopupElementScaleSettingForm />
     </Stack>
